+++ sdk/identity/azure-identity/azure/identity/_credentials/default.py
@@ -6,7 +6,7 @@ import logging
 import os
 
 from .._constants import EnvironmentVariables
-from .._internal import get_default_authority
+from .._internal import get_default_authority, normalize_authority
 from .browser import InteractiveBrowserCredential
 from .chained import ChainedTokenCredential
 from .environment import EnvironmentCredential
@@ -62,7 +62,8 @@ class DefaultAzureCredential(ChainedTokenCredential):
     """
 
     def __init__(self, **kwargs):
-        authority = kwargs.pop("authority", None) or get_default_authority()
+        authority = kwargs.pop("authority", None)
+        authority = normalize_authority(authority) if authority else get_default_authority()
 
         shared_cache_username = kwargs.pop("shared_cache_username", os.environ.get(EnvironmentVariables.AZURE_USERNAME))
         shared_cache_tenant_id = kwargs.pop(