+++ sdk/search/azure-search-documents/tests/async_tests/test_service_live_async.py
@@ -11,9 +11,9 @@ import time
 
 import pytest
 from azure.core.credentials import AzureKeyCredential
-from devtools_testutils import AzureMgmtTestCase, ResourceGroupPreparer
+from devtools_testutils import AzureMgmtTestCase
 
-from search_service_preparer import SearchServicePreparer
+from search_service_preparer import SearchServicePreparer, SearchResourceGroupPreparer
 
 from azure_devtools.scenario_tests.utilities import trim_kwargs_from_test_function
 
@@ -30,6 +30,9 @@ from azure.search.documents import(
     SearchServiceClient,
     ScoringProfile,
     Skillset,
+    DataSourceCredentials,
+    DataSource,
+    DataContainer
 )
 from azure.search.documents.aio import SearchServiceClient
 
@@ -37,6 +40,7 @@ CWD = dirname(realpath(__file__))
 SCHEMA = open(join(CWD, "..", "hotel_schema.json")).read()
 BATCH = json.load(open(join(CWD, "..", "hotel_small.json"), encoding='utf-8'))
 TIME_TO_SLEEP = 5
+CONNECTION_STRING = 'DefaultEndpointsProtocol=https;AccountName=storagename;AccountKey=NzhL3hKZbJBuJ2484dPTR+xF30kYaWSSCbs2BzLgVVI1woqeST/1IgqaLm6QAOTxtGvxctSNbIR/1hW8yH+bJg==;EndpointSuffix=core.windows.net'
 
 def await_prepared_test(test_fn):
     """Synchronous wrapper for async test methods. Used to avoid making changes
@@ -51,9 +55,8 @@ def await_prepared_test(test_fn):
 
     return run
 
-
-class SearchIndexClientTest(AzureMgmtTestCase):
-    @ResourceGroupPreparer(random_name_enabled=True)
+class SearchClientTest(AzureMgmtTestCase):
+    @SearchResourceGroupPreparer(random_name_enabled=True)
     @SearchServicePreparer()
     async def test_get_service_statistics(self, api_key, endpoint, **kwargs):
         client = SearchServiceClient(endpoint, AzureKeyCredential(api_key))
@@ -61,49 +64,56 @@ class SearchIndexClientTest(AzureMgmtTestCase):
         assert isinstance(result, dict)
         assert set(result.keys()) == {"counters", "limits"}
 
-    # Index operations
+class SearchIndexesClientTest(AzureMgmtTestCase):
 
-    @ResourceGroupPreparer(random_name_enabled=True)
+    @SearchResourceGroupPreparer(random_name_enabled=True)
     @SearchServicePreparer()
-    async def test_get_indexes_empty(self, api_key, endpoint, **kwargs):
-        client = SearchServiceClient(endpoint, AzureKeyCredential(api_key))
-        result = await client.get_indexes()
-        assert len(result) == 0
+    async def test_list_indexes_empty(self, api_key, endpoint, **kwargs):
+        client = SearchServiceClient(endpoint, AzureKeyCredential(api_key)).get_indexes_client()
+        result = await client.list_indexes()
 
-    @ResourceGroupPreparer(random_name_enabled=True)
+        with pytest.raises(StopAsyncIteration):
+            await result.__anext__()
+
+    @SearchResourceGroupPreparer(random_name_enabled=True)
     @SearchServicePreparer(schema=SCHEMA, index_batch=BATCH)
-    async def test_get_indexes(self, api_key, endpoint, index_name, **kwargs):
-        client = SearchServiceClient(endpoint, AzureKeyCredential(api_key))
-        result = await client.get_indexes()
-        assert len(result) == 1
-        assert result[0].name == index_name
+    async def test_list_indexes(self, api_key, endpoint, index_name, **kwargs):
+        client = SearchServiceClient(endpoint, AzureKeyCredential(api_key)).get_indexes_client()
+        result = await client.list_indexes()
 
-    @ResourceGroupPreparer(random_name_enabled=True)
+        first = await result.__anext__()
+        assert first.name == index_name
+
+        with pytest.raises(StopAsyncIteration):
+            await result.__anext__()
+
+    @SearchResourceGroupPreparer(random_name_enabled=True)
     @SearchServicePreparer(schema=SCHEMA, index_batch=BATCH)
     async def test_get_index(self, api_key, endpoint, index_name, **kwargs):
-        client = SearchServiceClient(endpoint, AzureKeyCredential(api_key))
+        client = SearchServiceClient(endpoint, AzureKeyCredential(api_key)).get_indexes_client()
         result = await client.get_index(index_name)
         assert result.name == index_name
 
-    @ResourceGroupPreparer(random_name_enabled=True)
+    @SearchResourceGroupPreparer(random_name_enabled=True)
     @SearchServicePreparer(schema=SCHEMA, index_batch=BATCH)
     async def test_get_index_statistics(self, api_key, endpoint, index_name, **kwargs):
-        client = SearchServiceClient(endpoint, AzureKeyCredential(api_key))
+        client = SearchServiceClient(endpoint, AzureKeyCredential(api_key)).get_indexes_client()
         result = await client.get_index_statistics(index_name)
         assert set(result.keys()) == {'document_count', 'storage_size'}
 
-    @ResourceGroupPreparer(random_name_enabled=True)
+    @SearchResourceGroupPreparer(random_name_enabled=True)
     @SearchServicePreparer(schema=SCHEMA, index_batch=BATCH)
     async def test_delete_indexes(self, api_key, endpoint, index_name, **kwargs):
-        client = SearchServiceClient(endpoint, AzureKeyCredential(api_key))
+        client = SearchServiceClient(endpoint, AzureKeyCredential(api_key)).get_indexes_client()
         await client.delete_index(index_name)
         import time
         if self.is_live:
             time.sleep(TIME_TO_SLEEP)
-        result = await client.get_indexes()
-        assert len(result) == 0
+        result = await client.list_indexes()
+        with pytest.raises(StopAsyncIteration):
+            await result.__anext__()
 
-    @ResourceGroupPreparer(random_name_enabled=True)
+    @SearchResourceGroupPreparer(random_name_enabled=True)
     @SearchServicePreparer(schema=SCHEMA, index_batch=BATCH)
     async def test_create_index(self, api_key, endpoint, index_name, **kwargs):
         name = "hotels"
@@ -129,14 +139,14 @@ class SearchIndexClientTest(AzureMgmtTestCase):
             fields=fields,
             scoring_profiles=scoring_profiles,
             cors_options=cors_options)
-        client = SearchServiceClient(endpoint, AzureKeyCredential(api_key))
+        client = SearchServiceClient(endpoint, AzureKeyCredential(api_key)).get_indexes_client()
         result = await client.create_index(index)
         assert result.name == "hotels"
         assert result.scoring_profiles[0].name == scoring_profile.name
         assert result.cors_options.allowed_origins == cors_options.allowed_origins
         assert result.cors_options.max_age_in_seconds == cors_options.max_age_in_seconds
 
-    @ResourceGroupPreparer(random_name_enabled=True)
+    @SearchResourceGroupPreparer(random_name_enabled=True)
     @SearchServicePreparer(schema=SCHEMA, index_batch=BATCH)
     async def test_create_or_update_index(self, api_key, endpoint, index_name, **kwargs):
         name = "hotels"
@@ -158,7 +168,7 @@ class SearchIndexClientTest(AzureMgmtTestCase):
             fields=fields,
             scoring_profiles=scoring_profiles,
             cors_options=cors_options)
-        client = SearchServiceClient(endpoint, AzureKeyCredential(api_key))
+        client = SearchServiceClient(endpoint, AzureKeyCredential(api_key)).get_indexes_client()
         result = await client.create_or_update_index(index_name=index.name, index=index)
         assert len(result.scoring_profiles) == 0
         assert result.cors_options.allowed_origins == cors_options.allowed_origins
@@ -178,20 +188,20 @@ class SearchIndexClientTest(AzureMgmtTestCase):
         assert result.cors_options.allowed_origins == cors_options.allowed_origins
         assert result.cors_options.max_age_in_seconds == cors_options.max_age_in_seconds
 
-    @ResourceGroupPreparer(random_name_enabled=True)
+    @SearchResourceGroupPreparer(random_name_enabled=True)
     @SearchServicePreparer(schema=SCHEMA, index_batch=BATCH)
     async def test_analyze_text(self, api_key, endpoint, index_name, **kwargs):
-        client = SearchServiceClient(endpoint, AzureKeyCredential(api_key))
+        client = SearchServiceClient(endpoint, AzureKeyCredential(api_key)).get_indexes_client()
         analyze_request = AnalyzeRequest(text="One's <two/>", analyzer="standard.lucene")
         result = await client.analyze_text(index_name, analyze_request)
         assert len(result.tokens) == 2
 
-    # Synonym Map operations
+class SearchSynonymMapsClientTest(AzureMgmtTestCase):
 
-    @ResourceGroupPreparer(random_name_enabled=True)
+    @SearchResourceGroupPreparer(random_name_enabled=True)
     @SearchServicePreparer(schema=SCHEMA, index_batch=BATCH)
     async def test_create_synonym_map(self, api_key, endpoint, index_name, **kwargs):
-        client = SearchServiceClient(endpoint, AzureKeyCredential(api_key))
+        client = SearchServiceClient(endpoint, AzureKeyCredential(api_key)).get_synonym_maps_client()
         result = await client.create_synonym_map("test-syn-map", [
             "USA, United States, United States of America",
             "Washington, Wash. => WA",
@@ -204,10 +214,10 @@ class SearchIndexClientTest(AzureMgmtTestCase):
         ]
         assert len(await client.get_synonym_maps()) == 1
 
-    @ResourceGroupPreparer(random_name_enabled=True)
+    @SearchResourceGroupPreparer(random_name_enabled=True)
     @SearchServicePreparer(schema=SCHEMA, index_batch=BATCH)
     async def test_delete_synonym_map(self, api_key, endpoint, index_name, **kwargs):
-        client = SearchServiceClient(endpoint, AzureKeyCredential(api_key))
+        client = SearchServiceClient(endpoint, AzureKeyCredential(api_key)).get_synonym_maps_client()
         result = await client.create_synonym_map("test-syn-map", [
             "USA, United States, United States of America",
             "Washington, Wash. => WA",
@@ -216,10 +226,10 @@ class SearchIndexClientTest(AzureMgmtTestCase):
         await client.delete_synonym_map("test-syn-map")
         assert len(await client.get_synonym_maps()) == 0
 
-    @ResourceGroupPreparer(random_name_enabled=True)
+    @SearchResourceGroupPreparer(random_name_enabled=True)
     @SearchServicePreparer(schema=SCHEMA, index_batch=BATCH)
     async def test_get_synonym_map(self, api_key, endpoint, index_name, **kwargs):
-        client = SearchServiceClient(endpoint, AzureKeyCredential(api_key))
+        client = SearchServiceClient(endpoint, AzureKeyCredential(api_key)).get_synonym_maps_client()
         await client.create_synonym_map("test-syn-map", [
             "USA, United States, United States of America",
             "Washington, Wash. => WA",
@@ -233,10 +243,10 @@ class SearchIndexClientTest(AzureMgmtTestCase):
             "Washington, Wash. => WA",
         ]
 
-    @ResourceGroupPreparer(random_name_enabled=True)
+    @SearchResourceGroupPreparer(random_name_enabled=True)
     @SearchServicePreparer(schema=SCHEMA, index_batch=BATCH)
     async def test_get_synonym_maps(self, api_key, endpoint, index_name, **kwargs):
-        client = SearchServiceClient(endpoint, AzureKeyCredential(api_key))
+        client = SearchServiceClient(endpoint, AzureKeyCredential(api_key)).get_synonym_maps_client()
         await client.create_synonym_map("test-syn-map-1", [
             "USA, United States, United States of America",
         ])
@@ -248,10 +258,10 @@ class SearchIndexClientTest(AzureMgmtTestCase):
         assert all(isinstance(x, dict) for x in result)
         assert set(x['name'] for x in result) == {"test-syn-map-1", "test-syn-map-2"}
 
-    @ResourceGroupPreparer(random_name_enabled=True)
+    @SearchResourceGroupPreparer(random_name_enabled=True)
     @SearchServicePreparer(schema=SCHEMA, index_batch=BATCH)
     async def test_create_or_update_synonym_map(self, api_key, endpoint, index_name, **kwargs):
-        client = SearchServiceClient(endpoint, AzureKeyCredential(api_key))
+        client = SearchServiceClient(endpoint, AzureKeyCredential(api_key)).get_synonym_maps_client()
         await client.create_synonym_map("test-syn-map", [
             "USA, United States, United States of America",
         ])
@@ -267,12 +277,12 @@ class SearchIndexClientTest(AzureMgmtTestCase):
             "Washington, Wash. => WA",
         ]
 
-    # Skillset operations
+class SearchDataSourcesClientTest(AzureMgmtTestCase):
 
-    @ResourceGroupPreparer(random_name_enabled=True)
+    @SearchResourceGroupPreparer(random_name_enabled=True)
     @SearchServicePreparer(schema=SCHEMA, index_batch=BATCH)
     async def test_create_skillset(self, api_key, endpoint, index_name, **kwargs):
-        client = SearchServiceClient(endpoint, AzureKeyCredential(api_key))
+        client = SearchServiceClient(endpoint, AzureKeyCredential(api_key)).get_skillsets_client()
 
         s = EntityRecognitionSkill(inputs=[InputFieldMappingEntry(name="text", source="/document/content")],
                                    outputs=[OutputFieldMappingEntry(name="organizations", target_name="organizations")])
@@ -287,10 +297,10 @@ class SearchIndexClientTest(AzureMgmtTestCase):
 
         assert len(await client.get_skillsets()) == 1
 
-    @ResourceGroupPreparer(random_name_enabled=True)
+    @SearchResourceGroupPreparer(random_name_enabled=True)
     @SearchServicePreparer(schema=SCHEMA, index_batch=BATCH)
     async def test_delete_skillset(self, api_key, endpoint, index_name, **kwargs):
-        client = SearchServiceClient(endpoint, AzureKeyCredential(api_key))
+        client = SearchServiceClient(endpoint, AzureKeyCredential(api_key)).get_skillsets_client()
         s = EntityRecognitionSkill(inputs=[InputFieldMappingEntry(name="text", source="/document/content")],
                                    outputs=[OutputFieldMappingEntry(name="organizations", target_name="organizations")])
 
@@ -302,10 +312,10 @@ class SearchIndexClientTest(AzureMgmtTestCase):
             time.sleep(TIME_TO_SLEEP)
         assert len(await client.get_skillsets()) == 0
 
-    @ResourceGroupPreparer(random_name_enabled=True)
+    @SearchResourceGroupPreparer(random_name_enabled=True)
     @SearchServicePreparer(schema=SCHEMA, index_batch=BATCH)
     async def test_get_skillset(self, api_key, endpoint, index_name, **kwargs):
-        client = SearchServiceClient(endpoint, AzureKeyCredential(api_key))
+        client = SearchServiceClient(endpoint, AzureKeyCredential(api_key)).get_skillsets_client()
         s = EntityRecognitionSkill(inputs=[InputFieldMappingEntry(name="text", source="/document/content")],
                                    outputs=[OutputFieldMappingEntry(name="organizations", target_name="organizations")])
 
@@ -320,10 +330,10 @@ class SearchIndexClientTest(AzureMgmtTestCase):
         assert len(result.skills) == 1
         assert isinstance(result.skills[0], EntityRecognitionSkill)
 
-    @ResourceGroupPreparer(random_name_enabled=True)
+    @SearchResourceGroupPreparer(random_name_enabled=True)
     @SearchServicePreparer(schema=SCHEMA, index_batch=BATCH)
     async def test_get_skillsets(self, api_key, endpoint, index_name, **kwargs):
-        client = SearchServiceClient(endpoint, AzureKeyCredential(api_key))
+        client = SearchServiceClient(endpoint, AzureKeyCredential(api_key)).get_skillsets_client()
         s = EntityRecognitionSkill(inputs=[InputFieldMappingEntry(name="text", source="/document/content")],
                                    outputs=[OutputFieldMappingEntry(name="organizations", target_name="organizations")])
 
@@ -334,10 +344,10 @@ class SearchIndexClientTest(AzureMgmtTestCase):
         assert all(isinstance(x, Skillset) for x in result)
         assert set(x.name for x in result) == {"test-ss-1", "test-ss-2"}
 
-    @ResourceGroupPreparer(random_name_enabled=True)
+    @SearchResourceGroupPreparer(random_name_enabled=True)
     @SearchServicePreparer(schema=SCHEMA, index_batch=BATCH)
     async def test_create_or_update_skillset(self, api_key, endpoint, index_name, **kwargs):
-        client = SearchServiceClient(endpoint, AzureKeyCredential(api_key))
+        client = SearchServiceClient(endpoint, AzureKeyCredential(api_key)).get_skillsets_client()
         s = EntityRecognitionSkill(inputs=[InputFieldMappingEntry(name="text", source="/document/content")],
                                    outputs=[OutputFieldMappingEntry(name="organizations", target_name="organizations")])
 
@@ -350,10 +360,10 @@ class SearchIndexClientTest(AzureMgmtTestCase):
         assert result.name == "test-ss"
         assert result.description == "desc2"
 
-    @ResourceGroupPreparer(random_name_enabled=True)
+    @SearchResourceGroupPreparer(random_name_enabled=True)
     @SearchServicePreparer(schema=SCHEMA, index_batch=BATCH)
     async def test_create_or_update_skillset_inplace(self, api_key, endpoint, index_name, **kwargs):
-        client = SearchServiceClient(endpoint, AzureKeyCredential(api_key))
+        client = SearchServiceClient(endpoint, AzureKeyCredential(api_key)).get_skillsets_client()
         s = EntityRecognitionSkill(inputs=[InputFieldMappingEntry(name="text", source="/document/content")],
                                    outputs=[OutputFieldMappingEntry(name="organizations", target_name="organizations")])
 
@@ -365,3 +375,70 @@ class SearchIndexClientTest(AzureMgmtTestCase):
         assert isinstance(result, Skillset)
         assert result.name == "test-ss"
         assert result.description == "desc2"
+
+class SearchDataSourcesClientTest(AzureMgmtTestCase):
+
+    def _create_datasource(self, name="sample-datasource"):
+        credentials = DataSourceCredentials(connection_string=CONNECTION_STRING)
+        container = DataContainer(name='searchcontainer')
+        data_source = DataSource(
+            name=name,
+            type="azureblob",
+            credentials=credentials,
+            container=container
+        )
+        return data_source
+
+    @SearchResourceGroupPreparer(random_name_enabled=True)
+    @SearchServicePreparer(schema=SCHEMA, index_batch=BATCH)
+    async def test_create_datasource_async(self, api_key, endpoint, index_name, **kwargs):
+        client = SearchServiceClient(endpoint, AzureKeyCredential(api_key)).get_datasources_client()
+        data_source = self._create_datasource()
+        result = await client.create_datasource(data_source)
+        assert result.name == "sample-datasource"
+        assert result.type == "azureblob"
+
+    @SearchResourceGroupPreparer(random_name_enabled=True)
+    @SearchServicePreparer(schema=SCHEMA, index_batch=BATCH)
+    async def test_delete_datasource_async(self, api_key, endpoint, index_name, **kwargs):
+        client = SearchServiceClient(endpoint, AzureKeyCredential(api_key)).get_datasources_client()
+        data_source = self._create_datasource()
+        result = await client.create_datasource(data_source)
+        assert len(await client.get_datasources()) == 1
+        await client.delete_datasource("sample-datasource")
+        assert len(await client.get_datasources()) == 0
+
+    @SearchResourceGroupPreparer(random_name_enabled=True)
+    @SearchServicePreparer(schema=SCHEMA, index_batch=BATCH)
+    async def test_get_datasource_async(self, api_key, endpoint, index_name, **kwargs):
+        client = SearchServiceClient(endpoint, AzureKeyCredential(api_key)).get_datasources_client()
+        data_source = self._create_datasource()
+        created = await client.create_datasource(data_source)
+        result = await client.get_datasource("sample-datasource")
+        assert result.name == "sample-datasource"
+
+    @SearchResourceGroupPreparer(random_name_enabled=True)
+    @SearchServicePreparer(schema=SCHEMA, index_batch=BATCH)
+    async def test_list_datasource_async(self, api_key, endpoint, index_name, **kwargs):
+        client = SearchServiceClient(endpoint, AzureKeyCredential(api_key)).get_datasources_client()
+        data_source1 = self._create_datasource()
+        data_source2 = self._create_datasource(name="another-sample")
+        created1 = await client.create_datasource(data_source1)
+        created2 = await client.create_datasource(data_source2)
+        result = await client.get_datasources()
+        assert isinstance(result, list)
+        assert set(x.name for x in result) == {"sample-datasource", "another-sample"}
+
+    @SearchResourceGroupPreparer(random_name_enabled=True)
+    @SearchServicePreparer(schema=SCHEMA, index_batch=BATCH)
+    async def test_create_or_update_datasource_async(self, api_key, endpoint, index_name, **kwargs):
+        client = SearchServiceClient(endpoint, AzureKeyCredential(api_key)).get_datasources_client()
+        data_source = self._create_datasource()
+        created = await client.create_datasource(data_source)
+        assert len(await client.get_datasources()) == 1
+        data_source.description = "updated"
+        await client.create_or_update_datasource(data_source)
+        assert len(await client.get_datasources()) == 1
+        result = await client.get_datasource("sample-datasource")
+        assert result.name == "sample-datasource"
+        assert result.description == "updated"