+++ sdk/servicebus/azure-servicebus/CHANGELOG.md
@@ -6,6 +6,7 @@
 
 * Added method `get_topic_sender` in `ServiceBusClient` to get a `ServiceBusSender` for a topic.
 * Added method `get_subscription_receiver` in `ServiceBusClient` to get a `ServiceBusReceiver` for a subscription under specific topic.
+* `ServiceBusSender.send()` can now send a list of messages in one call, if they fit into a single batch.  If they do not fit a `ValueError` is thrown.
 
 **BugFixes**
 
@@ -14,6 +15,8 @@
 **Breaking Changes**
 
 * Session receivers are now created via their own top level functions, e.g. `get_queue_sesison_receiver` and `get_subscription_session_receiver`.  Non session receivers no longer take session_id as a paramter.
+* `ServiceBusSender.send()` no longer takes a timeout parameter, as it should be redundant with retry options provided when creating the client.
+* Exception imports have been removed from module `azure.servicebus`. Import from `azure.servicebus.exceptions` instead.
 
 ## 7.0.0b1 (2020-04-06)